+++ src/main/java/com/furnitureapp/repository/distribution/DeliveryRepository.java
@@ -3,5 +3,5 @@ package com.furnitureapp.repository.distribution;
 import com.furnitureapp.entity.distribution.Delivery;
 import com.furnitureapp.repository.IRepository;
 
-public interface DeliveryRepository extends IRepository<Delivery, Long> {
+public interface DeliveryRepository extends IRepository<Delivery, Integer> {
 }
