+++ src/main/java/com/furnitureapp/repository/sales/CartRepository.java
@@ -3,5 +3,5 @@ package com.furnitureapp.repository.sales;
 import com.furnitureapp.entity.sales.Cart;
 import com.furnitureapp.repository.IRepository;
 
-public interface CartRepository extends IRepository<Cart, Long> {
+public interface CartRepository extends IRepository<Cart, Integer> {
 }
